// src/types/chat.ts
type BaseMessage<TType extends ChatMessageType> = {
  id: string
  type: TType
  content: ChatContentMap[TType]
  role: 'user' | 'assistant'
  timestamp: Date
  isStreaming?: boolean
}

/**
 * 채팅 메시지 타입 정의
 * role: 메시지를 보낸 주체 (사용자 또는 AI 어시스턴트)
 * content: 메시지 내용
 * timestamp: 메시지가 생성된 시간
 */
export type ChatTextMessage = BaseMessage<'text'>

export type ChatLLMAnswerStartMessage = BaseMessage<'llm_answer_start'>

export type ChatLLMAnswerMessage = BaseMessage<'llm_answer'>

export type ChatChartMessage = BaseMessage<'chart'>

export type ChatChartGroupMessage = BaseMessage<'chart_group'>

export type ChatTableMessage = BaseMessage<'table'>

export type ChatRecommendationQuestionMessage = BaseMessage<'recommendation_question'>

export type ChatMessage =
  | ChatTextMessage
  | ChatLLMAnswerStartMessage
  | ChatLLMAnswerMessage
  | ChatChartMessage
  | ChatChartGroupMessage
  | ChatTableMessage
  | ChatRecommendationQuestionMessage